Allow for averaged reconstructions
Time-averaged holdup is still interesting to inspect
but was hard to do with the codebase. This commit
updates loader functions to deal with large
datasets (>RAM) and load them as single frame, then
making average out of that.
Also introduces a time-averaged scan class - a combination
of fluidized bed and empty / full
This allows for reconstructing time-averaged holdup.