Improve scaling factor definition
Scaling factor was unclear to me. It seems to tie in to the density approximations as proposed for fluidized beds, somewhat combined with the gas fraction calculation.

Renamed it the density factor to make it a bit clearer that it has to do with density and thus gas fraction.

Re-implemented it calculation, so that every pixel gets a unique value now. That seems to make more sense to me, as each pixel is unique, and measuring a unique beam path.